엔티티 수정 후 광고 조회, 스트림 데이터 저장 수정

+++ src/main/java/com/sparta/binplay/dto/request/VideoRequestDto.java
@@ -9,9 +9,9 @@ public class VideoRequestDto {
     private Long videoId;
     private String title;
     private String description;
-    private long videoLength;
+    private int videoLength;
 
-    public VideoRequestDto(Long videoId, String title, String description, long videoLength) {
+    public VideoRequestDto(Long videoId, String title, String description, int videoLength) {
         this.videoId = videoId;
         this.title = title;
         this.description = description;
